分离robot，修正图片处理 & #29 fixed
