+++ lib/logger/logger.js
@@ -29,34 +29,33 @@ function wechatLogger(obj) {
 function logImageMessage(o, obj) {
   //debug("in webwxsync someone call me:" + inspect(o));
   // 查询用户名昵称
+  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${o.MsgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
+  // 保存图片到文件
+  var imgPath = path.join(process.cwd(), 'data/pic', crypto.createHash('md5').update(crypto.randomBytes(10)).digest('hex') + '.jpg');
+  try {
+    request.get(imgUrl, {jar: true}).pipe(fs.createWriteStream(imgPath));
+  } catch (e){
+    console.error('下载图像资源失败:', e);
+  }
   if (o.FromUserName.startsWith("@@")) {
-    logGroupImageMsg(o, obj);
+    logGroupImageMsg(o, obj, imgPath);
   } else {
-    logPrivateImageMsg(o, obj);
+    logPrivateImageMsg(o, obj, imgPath);
   }
 }
 
-function logPrivateImageMsg(o, obj) {
-  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${o.MsgId}&skey=${obj.BaseRequest.Skey}`;
-  var p = handlePrivate(o.FromUserName, imgUrl, obj);
+function logPrivateImageMsg(o, obj, imgPath) {
+  var p = handlePrivate(o.FromUserName, 'file://' + imgPath, obj);
   p.then(console.log, console.error);
 }
 
-function logGroupImageMsg(o, obj) {
+function logGroupImageMsg(o, obj, imgPath) {
   var result = /^(@[^:]+):<br\/>/mg.exec(o.Content);
   if (result) {
     var fromUserName = result[1];
   }
-  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${o.MsgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
-  // 保存图片到文件
-  var imgPath = path.join(process.cwd(), 'data/pic', crypto.createHash('md5').update(crypto.randomBytes(10)).digest('hex'));
-  try {
-    request.get(imgUrl, {jar: true}).pipe(fs.createWriteStream(imgPath));
-    var p = handleGroup(o.FromUserName, fromUserName + ':<br/>' + 'file://' + imgPath, obj);
-    p.then(console.log, console.error);
-  } catch (e){
-    console.error('下载图像资源失败:', e);
-  }
+  var p = handleGroup(o.FromUserName, fromUserName + ':<br/>' + 'file://' + imgPath, obj);
+  p.then(console.log, console.error);
 }
 
 function logTextMessage(o, obj) {
@@ -89,7 +88,6 @@ function handlePrivate(username, replyContent, obj) {
     contactP.then(_logPrivateTextMsg).catch(reject);
 
     function _logPrivateTextMsg(obj) {
-      // FIXME_TEST: 用find替换
       var m = obj.memberList.find(m=>m.UserName==username);
       resolve("[" + m.NickName + "说]" + replyContent);
     }
