+++ lib/webwx.js
@@ -235,12 +235,12 @@ function getContact(obj) {
   return p;
 }
 
-function botSpeak(obj) {
+function webwxsendmsg(obj) {
   if (!obj.webwxsync) {
     return Promise.resolve(obj);
   }
   var p = new Promise((resolve, reject)=>{
-    //debug('obj in botSpeak:\n' + inspect(obj));
+    //debug('obj in webwxsendmsg:\n' + inspect(obj));
     var BaseRequest = obj.BaseRequest;
     var pass_ticket = obj.pass_ticket;
 
@@ -266,8 +266,8 @@ function botSpeak(obj) {
         body: postData,
       };
 
-      //debug("options in botSpeak: \n" + inspect(options));
-      //debug("postData in botSpeak: \n" + inspect(postData));
+      //debug("options in webwxsendmsg: \n" + inspect(options));
+      //debug("postData in webwxsendmsg: \n" + inspect(postData));
 
       request(options, (error, response, body)=>{
         if (!error) {
@@ -385,7 +385,7 @@ function webwxsync(filters, mappers) {
         // get all replys resolved 所有回复完成
         // FIXME: webwxsendmsg似乎会被限制并发和频率，也可能只是微信的网络问题
         Promise.all(replys).then(()=>{
-          resolve(obj);   // 在回调中控制权交给botSpeak
+          resolve(obj);   // 在回调中控制权交给webwxsendmsg
         });
 
         // 更新联系人如果有的话
@@ -395,15 +395,6 @@ function webwxsync(filters, mappers) {
   }
 }
 
-function robot(filters, mappers) {
-  return (obj) => {
-    synccheck(obj)
-    .then(webwxsync(filters, mappers))
-    .then(botSpeak)
-    .then(robot(filters, mappers))
-    .catch(console.error);
-  }
-}
 
 function sessionStop(code, signal) {
   console.log('结束会话');
@@ -419,4 +410,6 @@ module.exports.login = login;
 module.exports.getbaseRequest = getbaseRequest;
 module.exports.webwxinit = webwxinit;
 module.exports.getContact = getContact;
-module.exports.robot = robot;
+module.exports.synccheck = synccheck;
+module.exports.webwxsync = webwxsync;
+module.exports.webwxsendmsg = webwxsendmsg;
